requirements.txt
assemblyai==0.31.0
google_api_python_client==2.122.0
gTTS==2.3.1
openai==1.40.2
pandas==2.2.2
protobuf==5.27.3
PyAudio==0.2.14
pydub==0.25.1
streamlit==1.33.0
